src/anomaly_detector/requirements.txt
kafka-python==2.0.2
scikit-learn
pandas
numpy
